editoast: move RouteCache to infra_cache
